# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: GramSathi AI API
  description: |
    AI-Powered Multilingual Government Assistant API for rural citizens in India.
    Enables scheme discovery, application processing, document OCR, grievance filing,
    notifications, and analytics across 22 Indian languages.
  version: "1.0.0"
  contact:
    name: GramSathi Team
    email: dev@gramsathi.ai
  license:
    name: Proprietary
servers:
  - url: https://api.gramsathi.ai/v1
    description: Production server
  - url: https://staging-api.gramsathi.ai/v1
    description: Staging server
  - url: http://localhost:8001/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token obtained from /auth/otp/verify
  schemas:
...
    UserOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mobile_number:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [citizen, admin, super_admin]
        is_verified:
          type: boolean
        preferred_language:
          type: string
          example: "hi"
        created_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        preferred_language:
          type: string
          maxLength: 10
        role:
          type: string
          enum: [citizen, admin, super_admin]
        is_active:
          type: boolean
    PaginatedUsers:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserOut"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer

    # Citizen Schemas
...
    GeoBreakdown:
      type: object
      properties:
        location:
          type: string
        total_citizens:
          type: integer
        total_applications:
          type: integer
        total_schemes:
          type: integer
        approval_rate:
          type: number
    GeoHeatmapPoint:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        intensity:
          type: number
        label:
          type: string
    AIMetrics:
      type: object
      properties:
        total_conversations:
          type: integer
        total_messages:
          type: integer
        avg_session_duration_seconds:
          type: number
        intent_distribution:
          type: object
        language_distribution:
          type: object
        user_satisfaction_score:
          type: number
    LanguageStats:
      type: object
      properties:
        language:
          type: string
        conversation_count:
          type: integer
        message_count:
          type: integer
        percentage:
          type: number
    ExportRequest:
      type: object
      required:
        - format
        - sections
      properties:
        format:
          type: string
          enum: [csv, json, xlsx]
        sections:
          type: array
          items:
            type: string
            enum: [overview, schemes, geo, ai]
        date_from:
          type: string
          format: date
          nullable: true
        date_to:
          type: string
          format: date
          nullable: true
    ExportResponse:
      type: object
      properties:
        export_id:
          type: string
        status:
          type: string
          enum: [processing, completed, failed]
        format:
          type: string
        download_url:
          type: string
          nullable: true
        expires_at:
          type: string
        created_at:
          type: string
    ExportStatus:
      type: object
      properties:
        export_id:
          type: string
        status:
          type: string
        progress:
          type: integer
        download_url:
          type: string
          nullable: true
        expires_at:
          type: string
        error:
          type: string
          nullable: true
...
paths:
...
  /admin/users:
    get:
      tags:
        - Admin Auth
      summary: List all users (admin)
      description: Returns paginated list of users with optional filters
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: is_verified
          in: query
          schema:
            type: boolean
        - name: role
          in: query
          schema:
            type: string
            enum: [citizen, admin, super_admin]
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or mobile number
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/users/{user_id}:
    get:
      tags:
        - Admin Auth
      summary: Get user by ID
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
    put:
      tags:
        - Admin Auth
      summary: Update user
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
    delete:
      tags:
        - Admin Auth
      summary: Deactivate user
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deactivated
  /admin/users/{user_id}/verify:
    post:
      tags:
        - Admin Auth
      summary: Verify user identity
      operationId: verifyUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User verified

  # ===== Citizen Service =====
...
  /analytics/geo:
    get:
      tags:
        - Geographic Analytics
      summary: Get geographic breakdown
      operationId: getGeoBreakdown
      security:
        - BearerAuth: []
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [state, district]
            default: district
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Geographic breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeoBreakdown"
  /analytics/geo/heatmap:
    get:
      tags:
        - Geographic Analytics
      summary: Get geographic heatmap data
      operationId: getGeoHeatmap
      security:
        - BearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Heatmap points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeoHeatmapPoint"
  /analytics/ai-performance:
    get:
      tags:
        - AI Analytics
      summary: Get AI performance metrics
      operationId: getAiPerformanceMetrics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: AI performance metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIMetrics"
  /analytics/ai/languages:
    get:
      tags:
        - AI Analytics
      summary: Get AI language usage stats
      operationId: getLanguageStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Language statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LanguageStats"
  /analytics/export:
    post:
      tags:
        - Analytics Export
      summary: Create analytics export job
      operationId: createExport
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: Export job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
  /analytics/exports/{export_id}:
    get:
      tags:
        - Analytics Export
      summary: Get export job status
      operationId: getExportStatus
      security:
        - BearerAuth: []
      parameters:
        - name: export_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Export status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportStatus"
  /analytics/exports/{export_id}/download:
    get:
      tags:
        - Analytics Export
      summary: Download completed export
      operationId: downloadExport
      security:
        - BearerAuth: []
      parameters:
        - name: export_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Download URL